requirements.txt
mock==3.0.5
python-dateutil==2.8.1
fixerio==1.0.0a0
pytest==5.3.4
flake8==3.7.9
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-RESTful==0.3.7
psycopg2